Got invitations and link join working, note for future: Users are created when a new invitation is created, so some users might not have all paramaters (for example, name)

+++ app/controllers/users/sessions_controller.rb
@@ -1,34 +1,14 @@
-# frozen_string_literal: true
-
 class Users::SessionsController < Devise::SessionsController
-  # before_action :configure_sign_in_params, only: [:create]
-
-  # GET /resource/sign_in
-  # def new
-  #   super
-  # end
-
-  # POST /resource/sign_in
   def create
-    super do |user|
+    super do |resource|
       if session[:invited_team_id]
-        team = Team.find(session[:invited_team_id])
-        resource.update(team: team, role: "member") if team
-
-        session.delete(:invited_team_id)  # Clear session after use
+        team = Team.find_by(id: session[:invited_team_id])
+        if team
+          resource.update(team: team, role: "member")
+          session.delete(:invited_team_id)
+          flash[:notice] = "Welcome to #{team.name}!"
+        end
       end
     end
   end
-
-  # DELETE /resource/sign_out
-  # def destroy
-  #   super
-  # end
-
-  # protected
-
-  # If you have extra params to permit, append them to the sanitizer.
-  # def configure_sign_in_params
-  #   devise_parameter_sanitizer.permit(:sign_in, keys: [:attribute])
-  # end
 end
